requirements.txt
PySide6
Pillow
opencv-contrib-python
rembg
transformers
torch
PyInstaller
